+++ Cargo.toml
@@ -5,8 +5,11 @@ members = ["crates/*"]
 [workspace.dependencies]
 chrono = "0.4.42"
 convert_case = "0.10.0"
+darling = "0.23.0"
 indexmap = "2.13.0"
 miette = "7.6.0"
+proc-macro2 = "1.0.95"
+quote = "1.0.40"
 regex = "1.12.2"
 relative-path = "2.0.1"
 reqwest = { version = "0.13.1", default-features = false }
@@ -19,6 +22,7 @@ serde_json = "1.0.149"
 serde_yaml = "0.9.33"
 serde_norway = "0.9.42"
 starbase_sandbox = "0.10.7"
+syn = "2.0.101"
 toml = "1.1.2"
 tracing = "0.1.44"
 url = "2.5.8"
