+++ crates/core/src/field_value.rs
@@ -1,0 +1,13 @@
+use syn::Type;
+
+pub enum WrapperType {
+    Arc,
+    Box,
+    Option,
+    Rc,
+}
+
+pub struct FieldValue {
+    pub original_ty: Type,
+    pub wrappers: Vec<WrapperType>,
+}
