+++ crates/core/src/lib.rs
@@ -1,0 +1,13 @@
+pub mod args;
+pub mod container;
+pub mod field;
+pub mod field_value;
+pub mod utils;
+pub mod variant;
+
+// #[cfg(feature = "config")]
+// pub mod config;
+// #[cfg(feature = "config")]
+// pub mod config_enum;
+#[cfg(feature = "schema")]
+pub mod schematic;
