+++ crates/macros-next/Cargo.toml
@@ -14,11 +14,9 @@ all-features = true
 proc-macro = true
 
 [dependencies]
-convert_case = "0.8.0"
-darling = "0.20.11"
-proc-macro2 = "1.0.95"
-quote = "1.0.40"
-syn = { version = "2.0.101", features = ["full"] }
+schematic_core = { version = "0.18.7", path = "../core" }
+quote = { workspace = true }
+syn = { workspace = true }
 
 [features]
 default = ["schema"]
